convertidor-imagenes-web/requirements.txt
Flask
Pillow
Werkzeug
redis
rq
PyJWT
psutil
kubernetes
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-redis